+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Update: Helipad (Podcasting 2.0 Boostagram reader) v0.2.0 [details](https://github.com/Podcastindex-org/helipad/releases/tag/v0.2.0)
 - Update: Mempool 3.0.0 [details](https://github.com/mempool/mempool/releases/tag/v3.0.0)
 - Update: Balance of Satoshis 19.3.4 (bos) [details](https://github.com/alexbosworth/balanceofsatoshis/blob/v19.3.4/CHANGELOG.md#1934)
+- Update: Labelbase 2.2.2 [details](https://x.com/labelbase_space)
 
 ## What's new in Version 1.11.2 of RaspiBlitz?